Move common fixtures to a conftest file
